+++ dev_plans/20250527_iterative_route_refactor_phase5.md
@@ -60,33 +60,33 @@ This completes the full development loop for the home page route. The next route
 
 ## Dev Loop 2: Refactoring the Recipe Extraction Page Route (`get_recipe_extraction_page` at `/recipes/extract`)
 
-- [ ] **1. Code Migration for the `/recipes/extract` Route:**
-    - [ ] **In `meal_planner/routers/pages.py`:**
-        - [ ] Add necessary imports (if not already present). Based on the function in `main.py`, this will likely include `Div` from `fasthtml.common` and `create_extraction_form` from `meal_planner.ui.extract_recipe`. (We already have `with_layout` and `rt`).
-        - [ ] Copy the `get_recipe_extraction_page()` function from `meal_planner/main.py` to `meal_planner/routers/pages.py`.
-    - [ ] **In `meal_planner/main.py`:**
-        - [ ] Delete the `get_recipe_extraction_page()` function definition.
-        - [ ] Review if imports like `create_extraction_form` from `meal_planner.ui.extract_recipe` or `Div` from `fasthtml.common` can be removed from `main.py` if no other routes in `main.py` use them.
+- [x] **1. Code Migration for the `/recipes/extract` Route:**
+    - [x] **In `meal_planner/routers/pages.py`:**
+        - [x] Add necessary imports (if not already present). Based on the function in `main.py`, this will likely include `Div` from `fasthtml.common` and `create_extraction_form` from `meal_planner.ui.extract_recipe`. (We already have `with_layout` and `rt`).
+        - [x] Copy the `get_recipe_extraction_page()` function from `meal_planner/main.py` to `meal_planner/routers/pages.py`.
+    - [x] **In `meal_planner/main.py`:**
+        - [x] Delete the `get_recipe_extraction_page()` function definition.
+        - [x] Review if imports like `create_extraction_form` from `meal_planner.ui.extract_recipe` or `Div` from `fasthtml.common` can be removed from `main.py` if no other routes in `main.py` use them. (Outcome: `create_extraction_form` removed, `Div` via `fasthtml.common.*` kept)
 
-- [ ] **2. Update and Verify Existing Tests (likely in `tests/test_main/test_route_endpoints.py`):**
-    - [ ] Identify the test(s) for the `/recipes/extract` endpoint (e.g., `test_get_recipe_extraction_page()`).
-    - [ ] Run the test suite (e.g., `./run_fast_coverage.sh`).
-    - [ ] Verify the identified test(s) pass.
-    - [ ] Confirm test coverage for the `/recipes/extract` route functionality (now in `meal_planner.routers.pages`) is maintained.
+- [x] **2. Update and Verify Existing Tests (likely in `tests/test_main/test_route_endpoints.py`):**
+    - [x] Identify the test(s) for the `/recipes/extract` endpoint (e.g., `test_get_recipe_extraction_page()`). (Outcome: No specific, isolated unit test found in the primary test file for this GET page route)
+    - [x] Run the test suite (e.g., `./run_fast_coverage.sh`).
+    - [x] Verify the identified test(s) pass.
+    - [x] Confirm test coverage for the `/recipes/extract` route functionality (now in `meal_planner.routers.pages`) is maintained.
 
-- [ ] **3. Move Tests to `tests/routers/test_pages.py`:**
-    - [ ] **In `tests/routers/test_pages.py`:**
-        - [ ] Add necessary imports (if not already present).
-        - [ ] Move the specific test function(s) for the `/recipes/extract` endpoint to this file.
-    - [ ] **In the source test file (e.g., `tests/test_main/test_route_endpoints.py`):**
-        - [ ] Delete the test function(s) that were moved.
+- [x] **3. Move Tests to `tests/routers/test_pages.py`:**
+    - [x] **In `tests/routers/test_pages.py`:**
+        - [x] Add necessary imports (if not already present).
+        - [x] Move the specific test function(s) for the `/recipes/extract` endpoint to this file. (N/A for this route as no specific test was moved)
+    - [x] **In the source test file (e.g., `tests/test_main/test_route_endpoints.py`):**
+        - [x] Delete the test function(s) that were moved. (N/A for this route)
 
-- [ ] **4. Confirm Tests and Coverage Post-Test-Move:**
-    - [ ] Run the test suite again (`./run_fast_coverage.sh`).
-    - [ ] Verify all tests pass.
-    - [ ] Verify test coverage is maintained, with the new functionality in `meal_planner/routers/pages.py` being tested by `tests/routers/test_pages.py`.
+- [x] **4. Confirm Tests and Coverage Post-Test-Move:**
+    - [x] Run the test suite again (`./run_fast_coverage.sh`).
+    - [x] Verify all tests pass.
+    - [x] Verify test coverage is maintained, with the new functionality in `meal_planner/routers/pages.py` being tested by `tests/routers/test_pages.py`. (N/A for specific test move, overall coverage maintained)
 
-- [ ] **5. Code Quality Checks & User Confirmation:**
-    - [ ] Run `uv run ruff format .`
-    - [ ] Run `uv run ruff check --fix .`
-    - [ ] Ask the user to confirm functionality and commit.
+- [x] **5. Code Quality Checks & User Confirmation:**
+    - [x] Run `uv run ruff format .`
+    - [x] Run `uv run ruff check --fix .`
+    - [x] Ask the user to confirm functionality and commit.
